resources/requirements.txt
# Dependencies for generate_icon.py
Pillow>=10.0.0
